Moved DataSourceWrapper to separate file.
