+++ thirdeye-core/src/main/java/org/apache/pinot/thirdeye/datasource/cache/DataSourceCache.java
@@ -142,22 +142,3 @@ public class DataSourceCache {
     cache.clear();
   }
 }
-
-class DataSourceWrapper {
-  private final ThirdEyeDataSource dataSource;
-  private final Timestamp updateTime;
-
-  public DataSourceWrapper(final ThirdEyeDataSource dataSource, final Timestamp updateTime) {
-    this.dataSource = dataSource;
-    this.updateTime = updateTime;
-  }
-
-  public ThirdEyeDataSource getDataSource() {
-    return dataSource;
-  }
-
-  public Timestamp getUpdateTime() {
-    return updateTime;
-  }
-}
-
